Simpler organization, same as for `PipelineType`
